requirements.txt
bintrees
pycrypto
bencode
miniupnpc
netifaces
simplejson
ipaddr
